FRONTEND: update: Acualizado los tab de enviar solicitud
