+++ src/hooks/useAdminUsuarios.js
@@ -10,16 +10,6 @@ export const useAdminUsuarios = () => {
   const [isLoading, setLoading] = useState(false);
   const [itemsAdminUsuarios, setAdminUsuarios] = useState([]);
   const [itemsRoles, setRoles] = useState([]);
-  const [itemsAgendarCitas, setAgendarCitas] = useState([]);
-  const [itemsHistorial, setHistorial] = useState([]);
-  const [itemsAprendices, setAprendices] = useState([]);
-  const [itemsSolicitudes, setConsultarSolicitud] = useState([]);
-  const [itemsSolicitudByID, setConsultarSolicitudByID] = useState([]);
-
-  
-
-  
-  
   //QUERY DE RESPUSTA DE CONSULTAS 
   const query = useCallback((itemUrl, tipo, opcion) => {
     setLoading(true);
@@ -50,21 +40,7 @@ export const useAdminUsuarios = () => {
                   case 'Roles':
                     setRoles(response)
                     break;
-                  case 'agendarCitas':
-                    setAgendarCitas(response) 
-                    break;
-                  case 'historial':
-                      setHistorial(response) 
-                   break; 
-                   case 'aprendices':
-                      setAprendices(response) 
-                   break;  
-                   case 'ConsultarSolicitud':
-                    setConsultarSolicitud(response) 
-                 break;                                    
-                 case 'ConsultarSolicitudByID':
-                  setConsultarSolicitudByID(response) 
-               break;                   
+                 
                 }
               })()
             }
@@ -87,11 +63,6 @@ export const useAdminUsuarios = () => {
       isLoading,
       itemsAdminUsuarios,
       itemsRoles,
-      itemsAgendarCitas,
-      itemsHistorial,
-      itemsAprendices,
-      itemsSolicitudes,
-      itemsSolicitudByID
     }
   )
 }
