+++ src/pages/dashboard/Project/ModuloSolicitudComite/ConsultarIncidente/CarSolicitudeEnviadas.js
@@ -2,25 +2,92 @@
 /* eslint-disable array-callback-return */
 /* eslint-disable react-hooks/exhaustive-deps */
 // @flow
-import React, { useContext } from 'react';
+import React, { useContext, useEffect, useState } from 'react';
 import { Row, Col, Card,  } from 'react-bootstrap';
-import { NotificacionesContext } from '../../../../../layouts/context/NotificacionesProvider';
-import { DashboardContext } from '../../../../../layouts/context/DashboardContext';
 //import Swal from 'sweetalert2';
 
  
 import Table from '../../../../../components/Table';
+import { DashboardContext } from '../../../../../layouts/context/DashboardContext';
+import { useAdminUsuarios } from '../../../../../hooks/useAdminUsuarios';
+import BtnSeccionAction from '../Components/BtnSeccionAction';
+import encodeBasicUrl from '../../../../../utils/encodeBasicUrl';
+import { NotificacionesContext } from '../../../../../layouts/context/NotificacionesProvider';
+import Swal from 'sweetalert2';
  
+const ActionColumn = ({ row }) => {
+
+  const {
+    eliminar,
+    validated,
+    toggle,
+    setOpen,
+    setItemsUpdate,
+    open, 
+    setitemsMenuPrincipal,setitemsUrl
+  } = useContext(DashboardContext);
+
+   const toggleSignUp = (id) => {
+    let permiso = sessionStorage.getItem('PERMISO');
+    const localPermiso = JSON.parse(permiso);
+    if (localPermiso?.update === 'N') {
+
+      if(row.cells[0].row.values.id===id)
+      setItemsUpdate(row?.cells[0]?.row?.values)
+      setOpen(open);
+      toggle()
+    } else {
+      Swal.fire('ESTA EN ESPERA DEL CONCEPTO DEL CLIENTE...');
+      /*
+      const url =`dashboard/ModuloSolicitudComite/EnviarSolicitud?id=${id}`
+      sessionStorage.setItem('ITEM_SELECT', JSON.stringify({ tipo: 'ConsultarIncidencia', menu: 'ModuloSolicitudComite'}));
+      setitemsMenuPrincipal('ModuloSolicitudComite');
+      setitemsUrl('ConsultarIncidencia');
+      setOpen(true);
+    return window.location.hash = url;
+    */
+    }
+  };
+
+  let permiso = sessionStorage.getItem('PERMISO');
+  const localPermiso = JSON.parse(permiso);
+  const obj = {
+    open,
+    toggleSignUp,
+    localPermiso,
+    validated,
+    key:row.cells[0].value,
+    row:row.cells[0].value,
+    eliminar,
+  }
+  return (
+    <React.Fragment>
+      <BtnSeccionAction obj={obj}> 
+      <Row>
+                <Col>
+                    <Card>
+                        <Card.Body>
+                            <h4 className="header-title mb-3">Conceptos</h4>
+                            <p className="text-muted font-14 mb-3">
+                            Hechos constitutivos de la presunta falta: El aprendiz se ausentó 4 días de los 5 días programados para la formación transversal correspondiente a la Competencia: Promover la interacción idónea consigo mismo, con los demás y con la naturaleza en los contextos laboral y social y RAPS: Asumir responsablemente los criterios de preservación y conservación del Medio Ambiente y de Desarrollo Sostenible, en el ejercicio de su desempeño laboral y social; cuyos motivos manifestados por el aprendiz sin presentar pruebas que soporten lo expresado, están registrados en el aplicativo Sofia Plus.
+                            </p>
+                        </Card.Body>
+                    </Card>
+                </Col>
+            </Row>   
+      </BtnSeccionAction>
+    </React.Fragment>
+  );
+};
 const CarSolicitudeEnviadas = (props) => {
   //const permisos = props.permisos || {};
-
  
-  const {itemsQueryByIdAprendiz} = useContext(NotificacionesContext)
+
+  const {itemsSolicitudByID} = useContext(NotificacionesContext)
   const {
     sizePerPageList
   } = useContext(DashboardContext);
-  const datos = itemsQueryByIdAprendiz?.data?.Solicitudes|| [{}];
- 
+  const datos = itemsSolicitudByID?.data?.Solicitudes|| [{}];
   const columns = [
     {
       Header: 'ID',
@@ -28,44 +95,27 @@ const CarSolicitudeEnviadas = (props) => {
       sort: true,
     },
     {
-      Header: 'codigoFicha',
-      accessor: 'codigoFicha',
-      sort: false,
-    },
-    {
       Header: 'Aprendiz',
       accessor: 'aprendiz',
       sort: true,
     },
     {
-      Header: 'Tipo Solicitud',
-      accessor: 'tipoSolicitud ',
+      Header: 'Estado',
+      accessor: 'estado',
       sort: true,
     }
     , {
-      Header: 'Tipo de Atención',
-      accessor: 'tipoAtencion',
-      sort: false,
-    },
-    {
-      Header: 'Fecha Hora Hechos',
+      Header: 'Fecha Incidencia',
       accessor: 'fechaHora',
       sort: false,
-    },
-    {
-      Header: 'Fecha Hora Propuesta',
-      accessor: 'fechaHoraPropuesta',
+    }, {
+      Header: 'Actas',
+      accessor: 'actas',
       sort: false,
     },
-      {
-        Header: 'Fecha Hora Agendada',
-        accessor: 'fechaHoraAgendada',
-        sort: false,
-      },
-      
-      {
-        Header: 'Estado',
-        accessor: 'estado',
+    {
+        Header: 'Tipo',
+        accessor: 'tipoAtencion',
         sort: false,
       },
     {
@@ -73,9 +123,9 @@ const CarSolicitudeEnviadas = (props) => {
       accessor: 'action',
       sort: false,
       classes: 'table-action',
+      Cell: ActionColumn,
     },
   ];
- 
   return (
     <>
 
@@ -94,8 +144,8 @@ const CarSolicitudeEnviadas = (props) => {
                     searchBoxClass="mt-2 mb-3"
                     isSearchable={true}
                     nametable={props.accion}
-                    titleTable={'HISTORIAL DE SOLICITUDES'}
-  />}
+                    titleTable={'HISTORIAL DE INCIDENCIAS'}
+              />}
             </Card.Body>
           </Card>
         </Col>
